refactor: restructured package organization
campus.tech.kakao.map
├── data
│   ├── db
│   │   ├── DataBase
│   │   ├── SearchHistory
│   │   ├── SearchHistoryDao
│   │   └── remote.api
│   │       └── dto
│   │           └── KakaoLocalAPI.kt
│   ├── di
│   │   └── AppModule
│   └── repository
│       └── KakaoRepository
├── ui
│   ├── activity
│   └── adapter
└── KyleMaps
