+++ app/src/main/AndroidManifest.xml
@@ -18,15 +18,15 @@
         tools:targetApi="31">
 
         <activity
-            android:name=".MapActivity"
+            android:name=".ui.activity.MapActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".AuthenticationErrorActivity" />
-        <activity android:name=".SearchActivity" />
+        <activity android:name=".ui.activity.AuthenticationErrorActivity" />
+        <activity android:name=".ui.activity.SearchActivity" />
 
     </application>
