+++ app/src/main/java/campus/tech/kakao/map/data/remote/api/dto/RegionInfo.kt
@@ -1,4 +1,4 @@
-package campus.tech.kakao.map.dto
+package campus.tech.kakao.map.data.remote.api.dto
 
 data class RegionInfo(
     val region: List<String>,
