+++ app/src/main/res/layout/activity_search.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:background="#FAF9F7"
     android:orientation="vertical"
-    tools:context=".SearchActivity">
+    tools:context=".ui.activity.SearchActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
